Unit_converter
A unit conversion expresses the same property as a different unit of measurement. For instance, time can be expressed in minutes instead of hours, while distance can be converted from miles to kilometers, or feet, or any other measure of length. Often measurements are given in one set of units, such as feet, but are needed in different units, such as chains. A conversion factor is a numeric expression that enables feet to be changed to chains as an equal exchange. A conversion factor is a number used to change one set of units to another, by multiplying or dividing. When a conversion is necessary, the appropriate conversion factor to an equal value must be used. For example, to convert inches to feet, the appropriate conversion value is 12 inches equal 1 foot. To convert minutes to hours, the appropriate conversion value is 60 minutes equal 1 hour.